+++ Cuda/Src/header/fillMatrix.h
@@ -3,8 +3,9 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include "coupleList.h"
 
-int **fillMatrix(int *premList, int sizePL, const Couple_List *R);
+int **fillMatrix(uint64_t *premList, int sizePL, const Couple_List *R);
 
 #endif
