+++ Cuda/Src/header/smooth.h
@@ -3,7 +3,8 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
-int isBSmoothG(int *primeList, int size, int y);
+int isBSmoothG(uint64_t *primeList, int size, uint64_t y);
 
 #endif
