+++ Cuda/Src/header/structure.h
@@ -9,6 +9,7 @@
 #define STRUCTURE_H_
 #include <cuda.h>
 #include <cuda_runtime.h>
+#include <stdint.h>
 
 struct couple {
 	int x;
@@ -26,6 +27,6 @@ struct cell {
 typedef struct cell* ensemble;
 
 __host__ __device__ ensemble initEns(int *size);
-__host__ __device__ int addCouple(ensemble *ens, int x, int y,int *size);
-__host__ __device__ int addVal(ensemble *ens, int x,int *size);
+__host__ __device__ int addCouple(ensemble *ens, uint64_t x, uint64_t y,int *size);
+__host__ __device__ int addVal(ensemble *ens, uint64_t x,int *size);
 #endif /* STRUCTURE_H_ */
